requirements.txt
fake_useragent==0.1.11
itemadapter==0.4.0
openpyxl==3.0.7
pandas==1.2.5
psycopg2==2.9.3
PyDispatcher==2.0.6
Scrapy==2.5.1
selenium==3.141.0
SQLAlchemy==1.4.22
w3lib==1.22.0
webdriver_manager==3.4.2
